requirements.txt
asgiref==3.8.1
bleach==6.1.0
distlib==0.3.8
Django==5.0.4
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
filelock==3.14.0
pillow==10.3.0
platformdirs==4.2.1
pydotplus==2.0.2
pyparsing==3.1.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.26.1
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
